requirements.txt
streamlit==1.22.0
langchain
python-dotenv==1.0.0
unstructured[local-inference]==0.9.2
sentence-transformers
faiss-cpu
openai
python-magic==0.4.27
langchain-openai
langchain-community
chromadb
joblib
scikit-learn

